feat: add service binding
